Read description for changes
Added a temporary healthbar for the Mothership's hull (Until Thomas has made the improved version).

Changed the color of the prototype Uranium to make it easier to see.

Fixed a BoD bug, where it could be fired again, after already having been fired.

Updated the asteroid prefabs to use the new sprites.

Made the Mothership more "floaty" allowing for movement even when shooting.

Made the scripts more manageable in the inspector via. headers and tooltips for certain references.

Switched spawn-locations for the Starfighter and Mothership, to make it easier for new players to figure out which ship they're controlling.

Fixed a bug where you would change turrets despite being in Thruster-Mode.

Made it possible to switch to the previous turret by pressing Arrow Down.