Read description for details
Added a sound effect for when shields are hit.

Added the blinking effect to turrets, Starfighter and Mothership hull when hit.

Fixed a bug where the top-right cannon's healthbar was darker than the others'.

Removed an unnecessary script that was used to control the thruster trail.

Placed the compass arrow in the left corner, made it red, and elongated it.