Read Description
Increased the speed at which the Mothership's turrets rotate by 50%.

Added an in-game pause menu that's accessible by pressing Escape (I'm aware of the bugs).

Made it so that the game actually stops when one of the players are killed, and a menu pops up where you're given the choice to quit to the main menu or do a rematch.

Applied the new sprites to the asteroids and "uranium" (Name still in progress).

Applied the new background that Bach made.