Read description
Implemented graphics for Rescue Slider.

Implemented graphics for Death Beam Slider.

Fixed an issue where you were able to shoot while the game was paused.

Made it so that the Starfighter bounces off of shielded objects (The Mothership and Planet).

Added the compass arrow to the rescue slider (Needs graphic)