Bouncy update (And other stuff)
Made it so that asteroids bounce off of the planet and mothership, and crystals bounce off of the planet.

Made it so that the Mothership bounces off of the planet.

Made it so that the game actually ends when the secondary win conditions are met (Beam of Death & Rescue System)