Read Description
Added some simple particles to the crystals.

Fixed a problem where the Starfighter would damage itself when being too close to an asteroid while shooting.

Rotated the background to make it fit the new player spawns.

Layers for the Mothership's art has been reordered.

Fixed an issue where max health for the Starfighter couldn't be changed in the inspector without messing up the healthbar.