Read Description
Reduced the time it takes to beam up an inhabitant and repairkit by 20%.

Added a sound for beaming up inhabitants and repairkits.

Increased drag and speed of players, allowing for more control.

Made it so that the Mothership can't shoot from outside the map.

Adjusted the volume of some sounds to make it easier to hear others.